Remove JsonSchema from nearly everywhere, tower lsp from errors
