+++ rust/kcl-lib/src/lint/rule.rs
@@ -1,12 +1,11 @@
 use anyhow::Result;
-use schemars::JsonSchema;
 use serde::Serialize;
 use tower_lsp::lsp_types::{Diagnostic, DiagnosticSeverity};
 
 use crate::{
     SourceRange,
     errors::Suggestion,
-    lsp::{IntoDiagnostic, to_lsp_edit},
+    lsp::{IntoDiagnostic, ToLspRange, to_lsp_edit},
     parsing::ast::types::{Node as AstNode, Program},
     walk::Node,
 };
@@ -30,7 +29,7 @@ where
 }
 
 /// Specific discovered lint rule Violation of a particular Finding.
-#[derive(Clone, Debug, ts_rs::TS, Serialize, JsonSchema)]
+#[derive(Clone, Debug, ts_rs::TS, Serialize)]
 #[ts(export)]
 #[cfg_attr(feature = "pyo3", pyo3::pyclass, pyo3_stub_gen::derive::gen_stub_pyclass)]
 #[serde(rename_all = "camelCase")]
@@ -125,7 +124,7 @@ impl IntoDiagnostic for &Discovered {
 }
 
 /// Abstract lint problem type.
-#[derive(Clone, Debug, PartialEq, ts_rs::TS, Serialize, JsonSchema)]
+#[derive(Clone, Debug, PartialEq, ts_rs::TS, Serialize)]
 #[ts(export)]
 #[cfg_attr(feature = "pyo3", pyo3::pyclass, pyo3_stub_gen::derive::gen_stub_pyclass)]
 #[serde(rename_all = "camelCase")]
