+++ rust/kcl-lib/src/lsp/mod.rs
@@ -9,7 +9,7 @@ pub mod test_util;
 mod tests;
 pub mod util;
 
-use tower_lsp::lsp_types::{Diagnostic, DiagnosticSeverity, DiagnosticTag};
+use tower_lsp::lsp_types::{Diagnostic, DiagnosticSeverity, DiagnosticTag, Position, Range};
 pub use util::IntoDiagnostic;
 
 use crate::{
@@ -55,4 +55,49 @@ pub type LspSuggestion = (Suggestion, tower_lsp::lsp_types::Range);
 pub(crate) fn to_lsp_edit(suggestion: &Suggestion, code: &str) -> LspSuggestion {
     let range = suggestion.source_range.to_lsp_range(code);
     (suggestion.clone(), range)
+}
+
+pub(crate) trait ToLspRange {
+    fn to_lsp_range(&self, code: &str) -> Range {
+        let start = self.start_to_lsp_position(code);
+        let end = self.end_to_lsp_position(code);
+        Range { start, end }
+    }
+
+    fn start_to_lsp_position(&self, code: &str) -> Position;
+    fn end_to_lsp_position(&self, code: &str) -> Position;
+}
+
+impl ToLspRange for crate::SourceRange {
+    fn start_to_lsp_position(&self, code: &str) -> Position {
+        // Calculate the line and column of the error from the source range.
+        // Lines are zero indexed in vscode so we need to subtract 1.
+        let mut line = code.get(..self.start()).unwrap_or_default().lines().count();
+        if line > 0 {
+            line = line.saturating_sub(1);
+        }
+        let column = code[..self.start()].lines().last().map(|l| l.len()).unwrap_or_default();
+
+        Position {
+            line: line as u32,
+            character: column as u32,
+        }
+    }
+
+    fn end_to_lsp_position(&self, code: &str) -> Position {
+        let lines = code.get(..self.end()).unwrap_or_default().lines();
+        if lines.clone().count() == 0 {
+            return Position { line: 0, character: 0 };
+        }
+
+        // Calculate the line and column of the error from the source range.
+        // Lines are zero indexed in vscode so we need to subtract 1.
+        let line = lines.clone().count() - 1;
+        let column = lines.last().map(|l| l.len()).unwrap_or_default();
+
+        Position {
+            line: line as u32,
+            character: column as u32,
+        }
+    }
 }
