+++ rust/kcl-lib/src/parsing/ast/types/condition.rs
@@ -1,4 +1,3 @@
-use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
 use super::{BoxNode, ConstraintLevel, Digest, Expr, Node, NodeList};
@@ -9,7 +8,7 @@ use crate::SourceRange;
 // https://github.com/KittyCAD/modeling-app/issues/4015
 type IfBlock = crate::parsing::ast::types::Program;
 
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, ts_rs::TS, JsonSchema)]
+#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, ts_rs::TS)]
 #[ts(export)]
 #[serde(tag = "type")]
 pub struct IfExpression {
@@ -21,7 +20,7 @@ pub struct IfExpression {
     pub digest: Option<Digest>,
 }
 
-#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, ts_rs::TS, JsonSchema)]
+#[derive(Debug, Clone, Deserialize, Serialize, PartialEq, ts_rs::TS)]
 #[ts(export)]
 #[serde(tag = "type")]
 pub struct ElseIf {
