+++ rust/kcl-lib/src/std/sweep.rs
@@ -3,7 +3,6 @@
 use anyhow::Result;
 use kcmc::{ModelingCmd, each_cmd as mcmd, length_unit::LengthUnit};
 use kittycad_modeling_cmds::{self as kcmc, shared::RelativeTo};
-use schemars::JsonSchema;
 use serde::Serialize;
 
 use super::{DEFAULT_TOLERANCE_MM, args::TyF64};
@@ -18,7 +17,7 @@ use crate::{
 };
 
 /// A path to sweep along.
-#[derive(Debug, Clone, Serialize, PartialEq, ts_rs::TS, JsonSchema)]
+#[derive(Debug, Clone, Serialize, PartialEq, ts_rs::TS)]
 #[ts(export)]
 #[serde(untagged)]
 #[allow(clippy::large_enum_variant)]
